+++ build.rs
@@ -8,6 +8,5 @@ fn main() {
         println!("cargo:rustc-link-lib=dylib=gfortran");
         println!("cargo:rustc-link-lib=dylib=openblas");
     } else {
-
     }
 }