+++ src/special/mod.rs
@@ -1,3 +1,4 @@
 //! Special function module
 
 pub mod function;
+pub mod legendre;